refactor(WishListService.java): Optional 사용해서 null값 처리하기

+++ src/main/java/gift/WishListService.java
@@ -1,7 +1,9 @@
 package gift;
 
 import org.springframework.stereotype.Service;
+
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class WishListService {
@@ -14,17 +16,29 @@ public class WishListService {
     }
 
     public List<WishList> getWishList(String email) {
-        Member member = memberRepository.findMemberByEmail(email);
+        Optional<Member> memberOptional = memberRepository.findMemberByEmail(email);
+        if (memberOptional.isEmpty()) {
+            throw new IllegalArgumentException("존재하지 않는 이메일입니다.");
+        }
+        Member member = memberOptional.get();
         return wishListRepository.findByMemberId(member.getId());
     }
 
     public void addProductToWishList(String email, Long productId) {
-        Member member = memberRepository.findMemberByEmail(email);
+        Optional<Member> memberOptional = memberRepository.findMemberByEmail(email);
+        if (memberOptional.isEmpty()) {
+            throw new IllegalArgumentException("존재하지 않는 이메일입니다.");
+        }
+        Member member = memberOptional.get();
         wishListRepository.addProductToWishList(member.getId(), productId);
     }
 
     public void removeProductFromWishList(String email, Long productId) {
-        Member member = memberRepository.findMemberByEmail(email);
+        Optional<Member> memberOptional = memberRepository.findMemberByEmail(email);
+        if (memberOptional.isEmpty()) {
+            throw new IllegalArgumentException("존재하지 않는 이메일입니다.");
+        }
+        Member member = memberOptional.get();
         wishListRepository.removeProductFromWishList(member.getId(), productId);
     }
 }
